Add constants and types files
