+++ src/pages/tokenLending/LoansTokenPage/TokenLoansContent/components/FilterSection.tsx
@@ -3,7 +3,7 @@ import { useState } from 'react'
 import { SearchSelect, SearchSelectProps } from '@banx/components/SearchSelect'
 import { SortDropdown, SortDropdownProps } from '@banx/components/SortDropdown'
 
-import { SortField } from '../hooks'
+import { SortField } from '../types'
 
 import styles from '../TokenLoansContent.module.less'
